+++ src/locales/en/not-found.js
@@ -1,0 +1,7 @@
+// Copyright (c) 2020 Antti Kivi
+// Licensed under the MIT License
+
+export default {
+  notFoundTitle: 'Page not found',
+  notFoundContent: 'This page doesn’t exist. You should return to the front page.',
+};