+++ src/locales/fi/not-found.js
@@ -1,0 +1,7 @@
+// Copyright (c) 2020 Antti Kivi
+// Licensed under the MIT License
+
+export default {
+  notFoundTitle: 'Sivua ei ole',
+  notFoundContent: 'Tätä sivua ei ole olemassa. Kannattaa palata etusivulle.',
+};